Generating Observed Networks - X

Given a network Y, we can generate N observed adjacency matrices as would have been reported by reporting nodes $m$ ($m in N$).

Usage

build_X(
  synthetic,
  mutuality = 0.5,
  sh_theta = 2,
  sc_theta = 0.5,
  theta = NULL,
  flag_self_reporter = T,
  Q = NULL,
  cutoff_X = F,
  lambda_diff = NULL,
  verbose = T,
  ...
)

Arguments

synthetic

A synthetic model

mutuality

The mutuality parameter (from 0 to 1)

sh_theta

Shape of gamma distribution from which to draw theta. The 'reliability' of nodes is represented by the parameter $theta_lm$ and by default are modelled as a gamma function with shape sh_theta and scale sc_theta.

sc_theta

Scale of gamma distribution from which to draw theta.

theta

Custom values of theta, if provided sh_theta and sc_theta will be ignored.

flag_self_reporter

Indicates whether a node can only report about their own ties (default is true).

Q

Maximum value of X entries. If None, it will use the network's K parameter

cutoff_X

Whether to set X as a binary

lambda_diff

The difference between each subsequent K

verbose

Provides additional details

...

Additional args of synthetic$build_X() method.

Value

A sptensor

Examples

random_net <- gm_Multitensor(N=100, M=100, L=1, eta=0.99)
#> Loaded vimure version 
X <- build_X(random_net, flag_self_reporter=TRUE, cutoff_X=FALSE, seed=10L)
dim(X)
#> [1]   1 100 100 100